Potential fix for code scanning alert no. 47: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to prevent overflow. We can achieve this by casting one of the operands to size_t before performing the multiplication. This will ensure that the multiplication is done using the larger type, and the result will be correctly passed to memset.

Specifically, we will cast mb_w to size_t before multiplying it by mb_h on line 1904.

+++ libwebp-1.0.3/src/enc/vp8l_enc.c
@@ -1901,7 +1901,7 @@ int VP8LEncodeImage(const WebPConfig* const config,
   if (picture->extra_info != NULL) {
     const int mb_w = (width + 15) >> 4;
     const int mb_h = (height + 15) >> 4;
-    memset(picture->extra_info, 0, mb_w * mb_h * sizeof(*picture->extra_info));
+    memset(picture->extra_info, 0, (size_t)mb_w * mb_h * sizeof(*picture->extra_info));
   }
 
  Error:
